+++ R/statuser-package.R
@@ -12,9 +12,10 @@
 #'   \item \code{\link{table2}}: like table(), showing variable names, and with proportions & chi2 built in
 #'   \item \code{\link{desc_var}}: Descriptive statistics for variables (optional, by group(s))
 #' }
-#' @section Custom Stats (new):
+#' @section Custom tools from papers by Uri Simonsohn:
 #' \itemize{
 #'   \item \code{\link{twolines}}: Two-lines test for U-shapes (Simonsohn 2018)
+#'   \item \code{\link{interprobe}}: Probe and visualize nonlinear interactions (Simonsohn 2024; Montealegre & Simonsohn 2026)
 #' }
 #'
 #' @section Graphing:
@@ -23,8 +24,8 @@
 #'   \item \code{\link{plot_cdf}}: Plot empirical cumulative distribution functions (optional, by group)
 #'   \item \code{\link{plot_density}}: Plot density functions (optional, by group)
 #'   \item \code{\link{plot_freq}}: Plot frequency of observed values (optional, by group)
+#'   \item \code{\link{plot_means}}: Barplot of means with confidence intervals and (optionally) tests
 #'   \item \code{\link{plot_gam}}: Plot fitted GAM values for a focal predictor
-#'   \item \code{\link{text2}}: like text() adding text-alignment and background color
 #' }
 #'
 #' @section Formatting:
@@ -36,9 +37,11 @@
 #'
 #' @section Miscellaneous:
 #' \itemize{
+#'   \item \code{\link{clear}}: Clear environment, console, and all graphics devices
 #'   \item \code{\link{list2}}: Like list(), but unnamed objects are automatically named
 #'   \item \code{\link{convert_to_sql}}: Convert CSV files to SQL INSERT statements
-#'   \item \code{\link{clear}}: Clear environment, console, and all graphics devices
+#'   \item \code{\link{text2}}: like text() adding text-alignment and background color
+
 #' }
 #'
 #' @author
@@ -47,7 +50,11 @@
 #' @references
 #' Simonsohn, U. (2018). Two lines: A valid alternative to the invalid testing of 
 #' U-shaped relationships with quadratic regressions. \emph{Advances in Methods and 
-#' Practices in Psychological Science}, 1(4), 538-555. 
+#' Practices in Psychological Science}, 1(4), 538-555. \doi{10.1177/2515245918805755}
+#'
+#' Simonsohn, U. (2024). Interacting with curves: How to validly test and probe interactions in the real (nonlinear) world. \emph{Advances in Methods and Practices in Psychological Science}, 7(1), Article 25152459231207787. \doi{10.1177/25152459231207787}
+#'
+#' Montealegre, D., & Simonsohn, U. (2026). \emph{Johnson Neyman 2.0} (working paper).
 #' 
 #' @seealso
 #' Useful links: